feat: split action.js
