+++ service/movieApi/src/action/getHandlerFileList.js
@@ -1,0 +1,9 @@
+export const getHandlerFileList = ({ handleFileList }) => {
+  return async (req, res) => {
+    const { requestId } = req.query
+    const handleResult = handleFileList({ requestId  })
+
+    res.json({ result: handleResult })
+  }
+}
+
