+++ service/movieApi/src/action/getHandlerLookupChatgptResponse.js
@@ -1,0 +1,10 @@
+export const getHandlerLookupChatgptResponse = ({ handleLookupChatgptResponse }) => {
+  return async (req, res) => {
+    const { requestId } = req.query
+
+    const handleResult = handleLookupChatgptResponse({ requestId })
+
+    res.json({ result: handleResult })
+  }
+}
+
