+++ service/movieApi/src/action/getHandlerLookupResponse.js
@@ -1,0 +1,10 @@
+export const getHandlerLookupResponse = ({ handleLookupResponse }) => {
+  return async (req, res) => {
+    const { requestId } = req.query
+
+    const handleResult = handleLookupResponse({ requestId })
+
+    res.json({ result: handleResult })
+  }
+}
+
